change-guard/functions/gemini_augmenter/requirements.txt
functions-framework==3.*
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
google-cloud-firestore==2.16.0
requests==2.32.3